+++ src/shellyDevice.realgen3.test.ts
@@ -492,7 +492,7 @@ describe('Shellies', () => {
     device.destroy();
   }, 30000);
 
-  test('create a gen 3 shelly2pmg3 device and update', async () => {
+  test('create a gen 3 shelly2pmg3 mode cover device and update', async () => {
     if (getMacAddress() !== address) return;
     device = await ShellyDevice.create(shelly, log, '192.168.1.166');
     expect(device).not.toBeUndefined();
@@ -509,7 +509,8 @@ describe('Shellies', () => {
     expect(device.hasUpdate).toBe(false);
     expect(device.firmware).toBe('1.4.99-2pmg3prod0-ge3db05c'); // firmwareGen2
     expect(device.auth).toBe(false);
-    expect(device.name).toBe('2PM Gen3');
+    expect(device.name).toBe('2PM Gen3 Cover');
+    expect(device.profile).toBe('cover');
     expect(device.username).toBe('admin');
     expect(device.password).toBe('tango');
 
@@ -517,9 +518,9 @@ describe('Shellies', () => {
 
     await device.saveDevicePayloads('temp');
 
-    expect(device.components.length).toBe(13);
-    expect(device.getComponentNames()).toStrictEqual(['Ble', 'Cloud', 'Input', 'MQTT', 'Switch', 'Sys', 'Sntp', 'WiFi', 'WS']);
-    expect(device.getComponentIds()).toStrictEqual(['ble', 'cloud', 'input:0', 'input:1', 'mqtt', 'switch:0', 'switch:1', 'sys', 'sntp', 'wifi_ap', 'wifi_sta', 'wifi_sta1', 'ws']);
+    expect(device.components.length).toBe(12);
+    expect(device.getComponentNames()).toStrictEqual(['Ble', 'Cloud', 'Cover', 'Input', 'MQTT', 'Sys', 'Sntp', 'WiFi', 'WS']);
+    expect(device.getComponentIds()).toStrictEqual(['ble', 'cloud', 'cover:0', 'input:0', 'input:1', 'mqtt', 'sys', 'sntp', 'wifi_ap', 'wifi_sta', 'wifi_sta1', 'ws']);
 
     expect(device.bthomeTrvs.size).toBe(0);
 
@@ -646,6 +647,95 @@ describe('Shellies', () => {
     expect(device.bthomeSensors.get('bthomesensor:217')?.name).toBe('Illuminance');
     expect(device.bthomeSensors.get('bthomesensor:217')?.sensorId).toBe(5);
     expect(device.bthomeSensors.get('bthomesensor:217')?.sensorIdx).toBe(0);
+
+    const cover = device.getComponent('cover:0');
+    expect(cover).not.toBeUndefined();
+    if (!cover) return;
+    expect(cover?.getValue('voltage')).toBeGreaterThan(200);
+    expect(cover?.hasProperty('apower')).toBe(true);
+    expect(cover?.hasProperty('current')).toBe(true);
+    expect(cover?.hasProperty('aenergy')).toBe(true);
+    expect(cover?.hasProperty('freq')).toBe(true);
+
+    expect(cover?.getValue('current_pos')).toBe(0);
+
+    // prettier-ignore
+    if (isCoverComponent(cover)) {
+      cover.Open();
+      // await waiter('Open()', () => { return cover.getValue('state') === 'opening'; }, true, 30000);
+      // await waiter('Open() II', () => { return cover.getValue('current_pos') === 100 || (cover.getValue('state') === 'stopped' || cover.getValue('state') === 'open'); }, true, 30000);
+      await waiter('Open() III', () => { return cover.getValue('current_pos') === 100; }, true, 30000);
+
+      cover.Close();
+      await wait(3000);
+      
+      cover.Stop();
+      // console.error(`Stop() I state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      // await waiter('Stop() I', () => { return cover.getValue('state') === 'stopped'; }, true, 30000);
+      // console.error(`Stop() II state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      await waiter('Stop() II', () => { return cover.getValue('current_pos') !== 100; }, true, 30000);
+
+      cover.Close();
+      // console.error(`Close() AI state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      // await waiter('Close() AI', () => { return cover.getValue('state') === 'closing'; }, true, 30000);
+      // console.error(`Close() AII state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      // await waiter('Close() AII', () => { return cover.getValue('current_pos') === 0 || (cover.getValue('state') === 'stopped' || cover.getValue('state') === 'closed'); }, true, 30000);
+      // console.error(`Close() AIII state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      await waiter('Close() AIII', () => { return cover.getValue('current_pos') === 0; }, true, 30000);
+
+      cover.GoToPosition(10);
+      // console.error(`GoToPosition(10) state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      // await waiter('GoToPosition(10)', () => { return cover.getValue('state') === 'opening'; }, true, 30000);
+      // console.error(`GoToPosition(10) I state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      // await waiter('GoToPosition(10) II', () => { return cover.getValue('state') === 'stopped'; }, true, 30000);
+      // console.error(`GoToPosition(10) II state ${cover.getValue('state')} pos ${cover.getValue('current_pos')}`);
+      await waiter('GoToPosition(10) III', () => { return cover.getValue('current_pos') === 10; }, true, 30000);
+
+      cover.Close();
+      // await waiter('Close() BI', () => { return cover.getValue('state') === 'closing'; }, true, 30000);
+      // await waiter('Close() BII', () => { return cover.getValue('current_pos') === 0 || (cover.getValue('state') === 'stopped' || cover.getValue('state') === 'closed'); }, true, 30000);
+      await waiter('Close() BIII', () => { return cover.getValue('current_pos') === 0; }, true, 30000);
+    }
+
+    expect(cover?.getValue('current_pos')).toBe(0);
+
+    shelly.removeDevice(device);
+    device.destroy();
+  }, 120000);
+
+  test('create a gen 3 shelly2pmg3 mode switch device and update', async () => {
+    if (getMacAddress() !== address) return;
+    device = await ShellyDevice.create(shelly, log, '192.168.1.172');
+    expect(device).not.toBeUndefined();
+    if (!device) return;
+    shelly.addDevice(device);
+    expect((device as any).wsClient).not.toBeUndefined();
+    (device as any).wsClient?.start();
+
+    expect(device.gen).toBe(3);
+    expect(device.host).toBe('192.168.1.172');
+    expect(device.model).toBe('S3SW-002P16EU');
+    expect(device.mac).toBe('8CBFEA9DE29C');
+    expect(device.id).toBe('shelly2pmg3-8CBFEA9DE29C');
+    expect(device.hasUpdate).toBe(false);
+    expect(device.firmware).toBe('1.4.99-2pmg3prod0-ge3db05c'); // firmwareGen2
+    expect(device.auth).toBe(false);
+    expect(device.name).toBe('2PM Gen3 Switch');
+    expect(device.profile).toBe('switch');
+    expect(device.username).toBe('admin');
+    expect(device.password).toBe('tango');
+
+    await device.fetchUpdate();
+
+    await device.saveDevicePayloads('temp');
+
+    expect(device.components.length).toBe(13);
+    expect(device.getComponentNames()).toStrictEqual(['Ble', 'Cloud', 'Input', 'MQTT', 'Switch', 'Sys', 'Sntp', 'WiFi', 'WS']);
+    expect(device.getComponentIds()).toStrictEqual(['ble', 'cloud', 'input:0', 'input:1', 'mqtt', 'switch:0', 'switch:1', 'sys', 'sntp', 'wifi_ap', 'wifi_sta', 'wifi_sta1', 'ws']);
+
+    expect(device.bthomeTrvs.size).toBe(0);
+    expect(device.bthomeDevices.size).toBe(0);
+    expect(device.bthomeSensors.size).toBe(0);
 
     const component = device.getComponent('switch:0');
     expect(component).not.toBeUndefined();